Backend/requirements.txt
# Web框架
fastapi==0.111.0
uvicorn==0.29.0
sse-starlette==1.8.2

# 状态机与智能体
langgraph==1.0.2
langchain-core==0.2.3

# 数据库
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic-sqlalchemy==0.0.9

# 数据验证
pydantic==2.6.4
pydantic-settings==2.0.3

# 网络请求
httpx==0.27.0
requests==2.32.3

# 工具类
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
tenacity==8.2.3  # 实现指数退避重试

# 数据库驱动
asyncpg==0.29.0  # PostgreSQL异步驱动

# 日志与监控
structlog==24.1.0
python-json-logger==2.0.7

# 容器化相关
gunicorn==22.0.0

# 测试相关
pytest==7.4.3